feat!: Bounce player off of walls and damage them
This is still a very basic implementation. I plan to adjust the damage
based on the player's speed when hitting the wall and also adjusting the
push back he receives.

There is currently one issue I am aware of:
When the player flies into the same direction as the wall is aligned,
then he can get essentially a speed boost (because he gets pushed
along the wall, while not getting pushed enough to be outside).
I will ignore this for now and see how bad this becomes.