restructure!: Split `shooting.rs` logic
Now there is shooting as a module with

- bullet
- rocket
- rocket_explosion

This is probably not ideal either, but it will do for now. In my
experience this is the best way to restructure: Simply restructure the
architecture in the way that seems best in the moment (and short term
future).

The reason is that you simply cannot know what and how you will develop
your project. The above approach also works really well in conjunction
with the rule to restructure on demand.