fix!: Reset rematch OnEnter GameState::InGame
This seems to fix the rematch issue (where on player rematches even
though the didn't press rematch). The issue was (and still might be!)
that the rematch was reset, however if the player was holding down the
rematch button then it would set the rematch state to true while the
state was transitioning.

I believe this is fixed now, if it isn't, then we would need to move the
reset rematch somewhere OnEnter GameState::GameOver

Closes #9.