fix!: Remove SpawnRocketExplosion Event
Events cannot be properly rolled back and so anything that is gameplay
relevant must not be an Event. The fact that we spawned rockets with an
Event meant that we ocassionally got desyncs through the rockets.

This is now fixed by using very simple dependency injection (which makes
the code more messy and less modular, but there aren't many clean ways
to achieve this, as this is an inheret problem of rolling back).