fix: Add Hash trait to RollbackSound
This fixes the infinite desync errors. My best guess as to why this
happens is that without the Hash component we are either

  - hashing something like the handle which is unique to each client?
  - not hashing it properly

But without the Hash trait it shouldn't be in the desync detection at
all (at least as far as I understand things). Either way, this fixes the
issue.