restructure!: Use AssetLoader for SpriteSheet
The explosion now uses the asset loader to add to the texture atlas. The
`rocket_explosion.rs` also received quite a bit of changes in order to
get rid of the rocket shooting desync, related to #22. The issue is the
despawning in the `animate_rocket_explosions()`, however I didn't fix it
yet. Further debugging is required.