feat!: Add MainMenu state and screen
Added `connection_ready` to `Ready` in `session.rs` because of the
following:

If we receive a seed *before* we actually establish the peer to peer
connection, then there is a chance that one peer will not receive the
confirmation seed (the 0 seed). To counter thisk, we simply wait until the
connection is established and receive the seed afterwards.