feat!: Modularize trail effect and add to rocket
The trails are now children of the spawning entity (player or rocket
currently). The way we are handling despawning of the trails is also
cleaner. We disable spawning of particles if the parent is dead and
unpause a kill timer which will despawn the trail effect once finished.

Partially addresses #27. Also relates to #25.