feat!: Load turn credentials into assets
This isn't safe! However I decided that I don't care too much. An
attacker that would steal the credentials would need to reverse engineer
the binary file (that I would release) in order to obtain the
credentials. I don't think that this is really easy to do, and extremely
unlikely in the first place.