feat: Add basic session stats (ping)
This is the first step towards #41. It's very basic, the next steps are
to use a vector to store the stats of all remote peers (this is
overkill, as we will only have two players anyways, but it's pretty easy
to do and will actually make things cleaner).

I will also create a UI you toggle with something like TAB to display
session stats for all peers.